dom/html/test/test_bug237071.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=237071
-->
<head>
  <title>Test for Bug 237071</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=237071">Mozilla Bug 237071</a>
<p id="display"></p>
<div id="content" >
  <ol id="theOL" start="22">
    <li id="foo" >should be 22</li>
    <li id="foo23">should be 23</li>
  </ol>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 237071 **/
is($('theOL').start, 22, "OL start attribute mapped to .start, not just text attribute");
</script>
</pre>
</body>
</html>
